+++ R/Filter.R
@@ -278,12 +278,13 @@ makeFilter(
 makeFilter(
   name = "linear.correlation",
   desc = "Pearson correlation between feature and target",
-  pkg  = "FSelector",
+  pkg  = "Rfast",
   supported.tasks = "regr",
   supported.features = "numerics",
   fun = function(task, nselect, ...) {
-    y = FSelector::linear.correlation(getTaskFormula(task), data = getTaskData(task))
-    setNames(y[["attr_importance"]], getTaskFeatureNames(task))
+    d = getTaskData(task, target.extra = TRUE)
+    y = Rfast::correls(d$target, d$data, type = "pearson")
+    setNames(abs(y[, "correlation"]), getTaskFeatureNames(task))
   }
 )
 
@@ -295,12 +296,13 @@ makeFilter(
 makeFilter(
   name = "rank.correlation",
   desc = "Spearman's correlation between feature and target",
-  pkg  = "FSelector",
+  pkg  = "Rfast",
   supported.tasks = "regr",
   supported.features = "numerics",
   fun = function(task, nselect, ...) {
-    y = FSelector::rank.correlation(getTaskFormula(task), data = getTaskData(task))
-    setNames(y[["attr_importance"]], getTaskFeatureNames(task))
+    d = getTaskData(task, target.extra = TRUE)
+    y = Rfast::correls(d$target, d$data, type = "spearman")
+    setNames(abs(y[, "correlation"]), getTaskFeatureNames(task))
   }
 )
 
@@ -486,16 +488,13 @@ univariate = makeFilter(
 makeFilter(
   name = "anova.test",
   desc = "ANOVA Test for binary and multiclass classification tasks",
-  pkg = character(0L),
+  pkg = "Rfast",
   supported.tasks = c("classif"),
   supported.features = c("numerics"),
   fun = function(task, nselect, ...) {
-    data = getTaskData(task)
-    sapply(getTaskFeatureNames(task), function(feat.name) {
-      f = as.formula(stri_paste(feat.name,"~",getTaskTargetNames(task)))
-      aov.t = aov(f, data = data)
-      summary(aov.t)[[1]][1,'F value']
-    })
+    d = getTaskData(task, target.extra = TRUE)
+    y = Rfast::ftests(as.matrix(d$data), as.numeric(d$target))
+    setNames(y[, "stat"], getTaskFeatureNames(task))
   }
 )