+++ tests/testthat/test_base_resample.R
@@ -65,7 +65,11 @@ test_that("resampling, predicting train set works", {
   m = setAggregation(mmce, train.mean)
   r = resample(lrn, multiclass.task, rdesc, measures = m)
   expect_false(is.na(r$aggr["mmce.train.mean"]))
-
+  expect_false(anyNA(r$pred$time))
+  expect_false(is.null(r$pred$predict.type))
+  expect_false(is.null(r$pred$threshold))
+  expect_equal(getTaskDescription(multiclass.task), r$pred$task.desc)
+  
   rdesc = makeResampleDesc("CV", iters = 2, predict = "both")
   lrn = makeLearner("classif.rpart")
   m1 = setAggregation(mmce, train.mean)
@@ -73,6 +77,12 @@ test_that("resampling, predicting train set works", {
   r = resample(lrn, multiclass.task, rdesc, measures = list(m1, m2))
   expect_false(is.na(r$aggr["mmce.train.mean"]))
   expect_false(is.na(r$aggr["mmce.test.mean"]))
+  expect_false(anyNA(r$pred$time))
+  expect_false(is.null(r$pred$predict.type))
+  expect_false(is.null(r$pred$threshold))
+  expect_equal(getTaskDescription(multiclass.task), r$pred$task.desc)
+  
+  
 
 })
 
@@ -135,4 +145,38 @@ test_that("resample has error messages when prediction fails", {
 
   configureMlr(on.learner.error = on.learner.error.saved)
   configureMlr(on.learner.warning = on.learner.warning.saved)
+})
+
+
+test_that("resample is extended by an additional measure", {
+  lrn = makeLearner("classif.rpart", predict.type = "prob")
+
+  # check if it works with test, both
+  # FIXME: add "train" after https://github.com/mlr-org/mlr/issues/1284 has been fixed
+  predict = c("test", "both")
+  # check if it works with different aggregation methods
+  aggr = list(test.mean, test.median, test.sd, test.range, test.join)
+  for (a in aggr) {
+    for (p in predict) {
+      rdesc = makeResampleDesc("CV", iter = 3, predict = p)
+      measures = list(mmce, ber, auc, brier)
+      # set aggregation method
+      measures = lapply(measures, setAggregation, a)
+      #if (p == "train") measures = lapply(measures, setAggregation, train.mean)
+      # create ResampleResult with all measures
+      res.all = resample(lrn, binaryclass.task, rdesc, measures)
+      # create ResampleResult with one measure and add the other ones afterwards
+      res = resample(lrn, binaryclass.task, rdesc, measures[[1]])
+      res.add = addRRMeasure(res, measures[-1])
+
+      # check if both ResampleResult objects are equal
+      expect_equal(res.all$measures.train, res.add$measures.train)
+      expect_equal(res.all$measures.test, res.add$measures.test)
+      expect_equal(res.all$aggr, res.add$aggr)
+    }
+  }
+
+  # keep.pred must be TRUE
+  res = resample(lrn, binaryclass.task, cv3, mmce, keep.pred = FALSE)
+  expect_error(addRRMeasure(res, auc), "keep.pred")
 })