+++ NEWS.md
@@ -1,5 +1,12 @@
 # mlr 2.10:
 
+## general
+* update to irace 2.0 -- there are algorithmic changes in irace that may affect
+  performance
+
+## Bug fixes
+* fixed bug in resample when using predict = "train" (issue #1284)
+
 ## functions - general
 * generateFilterValuesData: fixed a bug wrt feature ordering
 * imputeLearner: fixed a bug when data actually contained no NAs
@@ -29,6 +36,8 @@
   change any API behavior on the outside, though
 * generateHyperParsEffectData and plotHyperParsEffect now support more than 2
   hyperparameters
+* linear.correlation, rank.correlation, anova.test: use Rfast instead of
+  FSelector/custom implementation now, performance should be much better
 
 ## functions - new
 * filter: randomForest.importance
@@ -43,6 +52,10 @@
 * makeLearners: create multiple learners at once
 * getLearnerId, getLearnerType, getLearnerPredictType, getLearnerPackages
 * getLearnerParamSet, getLearnerParVals
+* getRRPredictionList
+* addRRMeasure
+* plotResiduals
+* getLearnerShortName
 
 ## functions - renamed
 * Renamed rf.importance filter (now deprecated) to randomForestSRC.var.rfsrc
@@ -56,17 +69,21 @@
   removed parameters "minprob", "pvalue", "randomsplits"
   as these are set internally and cannot be changed by the user
 * regr.GPfit: some more params for correlation kernel
-* classif.xgboost, regr.xgboost: can now properly handle NAs (property was missing and other
-  problems)
+* classif.xgboost, regr.xgboost: can now properly handle NAs (property was missing and other problems), added "colsample_bylevel" parameter
 
 ## learners - new
 * multilabel.cforest
 * surv.gbm
 * regr.cvglmnet
 
+## learner - removed
+* classif.randomForestSRCSyn, regr.randomForestSRCSyn: due to continued stability issues
+
 ## measures - new
 * ssr, qsr, lsr
 * rrse, rae, mape
+* kappa, wkappa
+* msle, rmsle
 
 # mlr 2.9: